+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/utils/IDBUtils.kt
@@ -177,7 +177,7 @@ interface IDBUtils {
         }
 
         val id = getLong(_ID)
-        var date = if (isAboveAndroidQ) {
+        val date = if (isAboveAndroidQ) {
             var tmpTime = getLong(DATE_TAKEN) / 1000
             if (tmpTime == 0L) {
                 tmpTime = getLong(DATE_ADDED)
@@ -662,7 +662,7 @@ interface IDBUtils {
         selectionArgs: Array<String>?,
         sortOrder: String?
     ): Cursor? {
-        fun log(logFunc: (log: String) -> Unit) {
+        fun log(logFunc: (log: String) -> Unit, cursor: Cursor?) {
             if (LogUtils.isLog) {
                 val sb = StringBuilder()
                 sb.appendLine("uri: $uri")
@@ -670,16 +670,20 @@ interface IDBUtils {
                 sb.appendLine("selection: $selection")
                 sb.appendLine("selectionArgs: ${selectionArgs?.joinToString(", ")}")
                 sb.appendLine("sortOrder: $sortOrder")
+                // format ? in selection and selectionArgs to display in log
+                val sql = selection?.replace("?", "%s")?.format(*selectionArgs ?: emptyArray())
+                sb.appendLine("sql: $sql")
+                sb.appendLine("cursor count: ${cursor?.count}")
                 logFunc(sb.toString())
             }
         }
 
         try {
             val cursor = query(uri, projection, selection, selectionArgs, sortOrder)
-            log(LogUtils::info)
+            log(LogUtils::info, cursor)
             return cursor
         } catch (e: Exception) {
-            log(LogUtils::error)
+            log(LogUtils::error, null)
             LogUtils.error("happen query error", e)
             throw e
         }
@@ -694,6 +698,36 @@ interface IDBUtils {
             return it?.count ?: 0
         }
     }
+
+    fun getAssetCount(
+        context: Context,
+        option: FilterOption,
+        requestType: Int,
+        galleryId: String,
+    ): Int {
+        val cr = context.contentResolver
+        val args = ArrayList<String>()
+        var where = option.makeWhere(requestType, args, false)
+
+        run {
+            val result = StringBuilder(where)
+            if (galleryId != PhotoManager.ALL_ID) {
+                if (result.trim().isNotEmpty()) {
+                    result.append(" AND ")
+                }
+                result.append("$BUCKET_ID = ?")
+                args.add(galleryId)
+            }
+
+            where = result.toString()
+        }
+
+        val order = option.orderByCondString()
+        cr.logQuery(allUri, arrayOf(_ID), where, args.toTypedArray(), order).use {
+            return it?.count ?: 0
+        }
+    }
+
 
     fun getAssetsByRange(
         context: Context,