+++ example/lib/widget/live_photos_widget.dart
@@ -4,6 +4,7 @@
 ///
 import 'package:flutter/material.dart';
 import 'package:photo_manager/photo_manager.dart';
+import 'package:photo_manager_image_provider/photo_manager_image_provider.dart';
 import 'package:video_player/video_player.dart';
 
 class LivePhotosWidget extends StatefulWidget {