+++ lib/photo_manager.dart
@@ -19,7 +19,6 @@ export 'src/filter/custom/order_by_item.dart';
 
 export 'src/internal/enums.dart';
 export 'src/internal/extensions.dart';
-export 'src/internal/image_provider.dart';
 export 'src/internal/plugin.dart' show PhotoManagerPlugin;
 export 'src/internal/progress_handler.dart';